Move test functions into aproprieate test files.
