+++ pkg/token/execCredentialPlugin_test.go
@@ -143,3 +143,62 @@ func setupMocks(t *testing.T) (*gomock.Controller, *mock_token.MockTokenCache, *
 
 	return ctrl, tokenCache, tokenProvider, pluginWriter
 }
+
+func TestDeviceloginAndNonInteractive(t *testing.T) {
+	testData := []struct {
+		name            string
+		execInfoEnvTest string
+		options         Options
+		expectedError   string
+	}{
+		{
+			name:            "KUBERNETES_EXEC_INFO.spec.interactive: false and login mode is devicelogin",
+			execInfoEnvTest: `{"kind":"ExecCredential","apiVersion":"client.authentication.k8s.io/v1beta1","spec":{"interactive":false}}`,
+			options: Options{
+				LoginMethod: DeviceCodeLogin,
+			},
+			expectedError: "devicelogin is not supported if interactiveMode is 'never'",
+		},
+	}
+
+	for _, data := range testData {
+		t.Run(data.name, func(t *testing.T) {
+			os.Setenv("KUBERNETES_EXEC_INFO", data.execInfoEnvTest)
+			defer os.Unsetenv("KUBERNETES_EXEC_INFO")
+			ecp, err := New(&data.options)
+			if ecp != nil || err == nil || err.Error() != data.expectedError {
+				t.Fatalf("expected: return defined error, actual: did not return expected error")
+			}
+		})
+	}
+}
+
+func TestKUBERNETES_EXEC_INFOIsEmpty(t *testing.T) {
+	testData := []struct {
+		name            string
+		execInfoEnvTest string
+		options         Options
+	}{
+		{
+			name:            "KUBERNETES_EXEC_INFO is empty",
+			execInfoEnvTest: "",
+			options: Options{
+				LoginMethod: DeviceCodeLogin,
+				ClientID:    "clientID",
+				ServerID:    "serverID",
+				TenantID:    "tenantID",
+			},
+		},
+	}
+
+	for _, data := range testData {
+		t.Run(data.name, func(t *testing.T) {
+			os.Setenv("KUBERNETES_EXEC_INFO", data.execInfoEnvTest)
+			defer os.Unsetenv("KUBERNETES_EXEC_INFO")
+			ecp, err := New(&data.options)
+			if ecp == nil || err != nil {
+				t.Fatalf("expected: return execCredentialPlugin and nil error, actual: did not return execCredentialPlugin or did not return expected error")
+			}
+		})
+	}
+}
